Refactor SSR scripts
